Codde Fixing an Refactoring
